Create a queries directory and break apart queries into their specific type file
